+++ src/components/Widgets/Form/ValidatedField.jsx
@@ -1,8 +1,8 @@
 import React, { Component } from 'react'
 import PropTypes from 'prop-types'
 import styled from 'styled-components'
-import Checkmark from './AnimatedIcons/Checkmark'
-import Cross from './AnimatedIcons/AnimatedCross'
+import Checkmark from '../AnimatedIcons/Checkmark'
+import Cross from '../AnimatedIcons/AnimatedCross'
 
 const height = 50
